vistes de forma rara

+++ database/migrations/2024_03_05_125803_enemys.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name')->unique();
             $table->string('type');
             $table->integer('health');
-            $table->integer('damage');
+            $table->float('damage');
             $table->timestamps();
         });
     }
